Add documentation, fix command prompts and make plottype choice work
A bit too many changes in this commit.. Added documentation for the
whole module and did some code cleanup (mostly print statements that
were used for testing which you don't want to see when running the
program in the console).

The 'choose plottype'-prompt did nothing before, but now it actually
applies the choice you make.